Consumer controls work, let's focus on BTN6 next

+++ code/include/board.h
@@ -36,26 +36,26 @@
 
 /**
  * BTN0  - flash
- * BTN1  - volume mute
+ * BTN1  - camera toggle
  * BTN2  - volume increment
  * BTN3  - volume decrement
- * BTN4  - volume mute (?)
- * BTN5  - mute toggle
- * BTN6  - Hold to unmute
+ * BTN4  - volume mute
+ * BTN5  - microphone mute toggle
+ * BTN6  - Hold to unmute microphone
  * 
  * PIND0 - 
- * PIND1 - hold to mute
- * PIND2 - volume mute (?)
- * PIND3 - mute toggle 
+ * PIND1 - hold to unmute microphone
+ * PIND2 - volume mute
+ * PIND3 - microphone mute toggle 
  * PIND4 - volume decrement
  * PIND5 - volume increment
  * PIND6 - flash
- * PIND7 - volume mute 
+ * PIND7 - camera toggle 
  */
 
 #define LED PORTF0
 
-int8_t check_buttons(uint8_t* dout);
+int8_t check_buttons(uint16_t* dout);
 int8_t animate_led(uint8_t* din);
 
 #endif
